askmydocs-backend/requirements.txt
# FastAPI and related
fastapi>=0.95.0,<0.112.0
uvicorn>=0.22.0,<0.23
sqlalchemy>=2.0.22
asyncpg>=0.27.0
alembic>=1.11.1
python-multipart>=0.0.8

# LangChain & AI
langchain==0.3.7
langchain-openai==0.3.7
chromadb==1.1.1
openai>=1.58.1,<2.0.0

# Pydantic v2 settings

pydantic>=2.12.0,<3.0
pydantic-settings>=2.11.0,<3.0

httpx
pypdf
langchain-text-splitters
ollama>=0.0.1
reportlab>=4.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
requests>=2.25.0




